Fix: Resolver conflictos de redefinición C++ (uint8_t vs uint32_t) en omega_effect

+++ app/src/main/cpp/audio_effect_compat.h
@@ -175,7 +175,7 @@ static const effect_uuid_t EFFECT_UUID_NULL_VAL = {0, 0, 0, 0, {0,0,0,0,0,0}};
 #endif
 
 typedef uint32_t audio_channel_mask_t;
-typedef uint8_t  audio_format_t;
+typedef uint32_t audio_format_t;
 
 static inline int audio_channel_count_from_out_mask(uint32_t mask) {
     int n = 0;
